restructured project, implemented right rotate
